+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 - Fixed and `parseEvery` and `nextEvent`, more tests cases
 - Add year support to `Every` format
 - Add function `flatObject(sourceObject: object, fieldNames: Array<string>): object`
-- Add function `isObject(o: string | number | boolean | object ): boolean`
+- Add function `isHashObject(o: string | number | boolean | object): boolean`
 
 ## [3.5.21][] - 2022-06-27